services/frontend/requirements.txt
Flask==2.0.1
gunicorn==20.1.0
python-dotenv==0.20.0
requests==2.27.1

